Changes are made to the controller, service and user administrator test.
