+++ app/constants.js
@@ -1,3 +1,5 @@
 exports.PREVIOUS_MESSAGE = 'Making request';
 exports.MESSAGE_OK = 'Request made correctly, sending...';
 exports.MESSAGE_ALBUM_API_FAILED = 'Failed to consume the album API';
+exports.ADMIN_ROLE = 'admin';
+exports.REGULAR_ROLE = 'regular';
