+++ docs/admin.html
@@ -1,5 +1,5 @@
 <html><head><link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" /><script src="https://code.jquery.com/jquery-1.12.2.min.js"></script><script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script><style> pre { width: 95%; } </style></head><body><div class="container"><div class="jumbotron"><h1>Admin</h1></div><div class="panel panel-default"><div class="panel-heading" role="tab" id="heading-admin-users-POST"><h4 class="panel-title"><a role="button" data-toggle="collapse" data-parent="#accordion" aria-expanded="true" aria-controls="collapse-admin-users-POST" href="#collapse-admin-users-POST">POST /admin/users</a></h4></div><div id="collapse-admin-users-POST" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading-admin-users-POST"><div class="panel-body"><h4>Request Data</h4><div class="container"><h5>Headers</h5><pre>{
-  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTUzODI3fQ.9RryUZzI4BzMei4ShepLDw6SWh76BbgRnFqNpGOykWk",
+  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTcwNzc2fQ.jEdXzjmIkxnl1ofoIH0KISAR71tWgwunjPlI7-GCvXI",
   "content-type": "application/json",
   "content-length": 96
 }</pre><h5>Body</h5><pre>{
@@ -8,14 +8,14 @@
   "email": "jose.perez@wolox.com.ar",
   "password": "12345test"
 }</pre></div></div><div class="panel-body"><h4>Response Data</h4><div class="container"><h5>Status: 201</h5><h5>Headers</h5><pre>{
-  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTUzODI3fQ.9RryUZzI4BzMei4ShepLDw6SWh76BbgRnFqNpGOykWk",
+  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTcwNzc2fQ.jEdXzjmIkxnl1ofoIH0KISAR71tWgwunjPlI7-GCvXI",
   "content-type": "application/json",
   "content-length": 96
 }</pre><h5>Body</h5><pre>{
   "firstName": "Jose",
   "lastName": "Perez"
 }</pre></div></div></div></div><div class="panel panel-default"><div class="panel-heading" role="tab" id="heading-admin-users-POST"><h4 class="panel-title"><a role="button" data-toggle="collapse" data-parent="#accordion" aria-expanded="true" aria-controls="collapse-admin-users-POST" href="#collapse-admin-users-POST">POST /admin/users</a></h4></div><div id="collapse-admin-users-POST" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading-admin-users-POST"><div class="panel-body"><h4>Request Data</h4><div class="container"><h5>Headers</h5><pre>{
-  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTUzODI3fQ.9RryUZzI4BzMei4ShepLDw6SWh76BbgRnFqNpGOykWk",
+  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTcwNzc2fQ.jEdXzjmIkxnl1ofoIH0KISAR71tWgwunjPlI7-GCvXI",
   "content-type": "application/json",
   "content-length": 95
 }</pre><h5>Body</h5><pre>{
@@ -24,9 +24,9 @@
   "email": "jose.perez@wolox.com.ar",
   "password": "12345678"
 }</pre></div></div><div class="panel-body"><h4>Response Data</h4><div class="container"><h5>Status: 201</h5><h5>Headers</h5><pre>{
-  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTUzODI3fQ.9RryUZzI4BzMei4ShepLDw6SWh76BbgRnFqNpGOykWk",
+  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZmlyc3ROYW1lIjoiSm9obiIsImxhc3ROYW1lIjoiS2F0emVuYmFjaCIsImVtYWlsIjoiam9obi5rYXR6QHdvbG94LmNvIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNTYyOTcwNzc2fQ.jEdXzjmIkxnl1ofoIH0KISAR71tWgwunjPlI7-GCvXI",
   "content-type": "application/json",
   "content-length": 95
 }</pre><h5>Body</h5><pre>{
-  "message": "User Jose updated to admin"
+  "message": "User jose.perez@wolox.com.ar updated to admin"
 }</pre></div></div></div></div></div></body></html>
